+++ app/system/app/components/input-video.vue
@@ -83,26 +83,16 @@
                 this.$set('image', undefined);
                 this.$set('video', undefined);
 
-                if (matches = (src.match(/(?:\/\/.*?youtube\.[a-z]+)\/watch\?v=([^&]+)&?(.*)/) || src.match(/youtu\.be\/(.*)/))) {
+                if (matches = (src.match(/.*(?:youtu.be\/|v\/|u\/\w\/|embed\/|watch\?v=)([^#\&\?]*).*/))) {
 
                     this.image = '//img.youtube.com/vi/' + matches[1] + '/hqdefault.jpg';
 
-                } else if (src.match(/(\/\/.*?)vimeo\.[a-z]+\/([0-9]+).*?/)) {
+                } else if (src.match(/https?:\/\/(?:www\.|player\.)?vimeo.com\/(?:channels\/(?:\w+\/)?|groups\/([^\/]*)\/videos\/|album\/(\d+)\/video\/|video\/|)(\d+)(?:$|\/|\?)/)) {
 
-                    var id = btoa(src);
-
-                    if (this.$session[id]) {
-
-                        this.image = this.$session[id];
-
-                    } else {
-
-                        this.$http.get('http://vimeo.com/api/oembed.json', {url: src}).then(function (res) {
-                            var data = res.data;
-                            this.image = this.$session[id] = data.thumbnail_url;
-                        });
-
-                    }
+                    this.$http.get('http://vimeo.com/api/oembed.json', {url: src}, {cache: 10}).then(function (res) {
+                        var data = res.data;
+                        this.image = data.thumbnail_url;
+                    });
 
                 } else {