+++ app/system/modules/editor/app/components/link-preview.vue
@@ -1,6 +1,6 @@
 <template>
 
-    <a :class="link.class" :href="link.link" @click.prevent="$parent.openModal(link)">{{{ link.txt ? link.txt : 'Select Link' | trans }}} <span class="pk-icon-link pk-icon-hover"></span></a>
+    <a :class="link.class" :href="link.link" @click.prevent="$parent.openModal(link)"><span v-html="link.txt" v-if="link.txt"></span><span v-else>{{ 'Select Link' | trans }}</span> <span class="pk-icon-link pk-icon-hover"></span></a>
 
 </template>