+++ app/system/modules/user/app/views/registration.js
@@ -10,12 +10,15 @@ module.exports = {
     methods: {
 
         submit: function () {
+
+            this.error = null;
+
             this.$http.post('user/registration/register', {user: this.user}).then(function (res) {
-                        window.location.replace(res.data.redirect);
-                    }, function (error) {
-                        this.error = error.data;
-                    }
-                );
+                    window.location.replace(res.data.redirect);
+            }, function (error) {
+                    this.error = error.data;
+                }
+            );
         }
 
     }