+++ app/system/modules/user/views/message.php
@@ -1,0 +1,9 @@
+<p class="uk-alert <?= $success ? 'uk-alert-success' : 'uk-alert-danger'; ?>">
+    <?= $message ?>
+</p>
+
+<?php if ($link): ?>
+<div class="uk-margin-top uk-text-center">
+    <a href="<?= $link ?>" class="uk-button"><?= $label ?></a>
+</div>
+<?php endif; ?>