+++ app/system/modules/view/index.php
@@ -41,14 +41,8 @@ return [
 
         'site' => function ($event, $app) {
             $app->on('view.meta', function ($event, $meta) use ($app) {
-
-                $route = $app['url']->get($app['request']->attributes->get('_route'), $app['request']->attributes->get('_route_params', []), true);
-
-                if ($route != $app['request']->getSchemeAndHttpHost().$app['request']->getRequestUri()) {
-                    $meta->add('canonical', $route);
-                }
-
-            });
+                $meta->add('canonical', $app['url']->get($app['request']->attributes->get('_route'), $app['request']->attributes->get('_route_params', []), 0));
+            }, 60);
         },
 
         'view.init' => [function ($event, $view) {