+++ CHANGELOG.md
@@ -3,9 +3,15 @@
 ## WIP
 
 ### Added
- - Added redirect after login to user settings
- - Added view.init event
- - Added global params object to view
+- Added redirect after login to user settings
+- Added view.init event
+- Added global params object to view
+- Added file picker
+- Added support for script tag attributes 'defer' and 'async'
+ 
+### Changed
+- Transfer widget and menu positions on theme change
+- Image-, video-, link-picker: Preserve existing attributes
 
 ## 0.10.4 (March 1 2016)